requirements.txt
Flask==2.0.2
requests==2.11.0
newspaper3k==0.2.8
nltk==3.3.0
lxml-html-clean==0.1.0
